started generating unique urls and adding to db
